replica handling added to wiggler processing

+++ segway/run_segway.sh
@@ -17,8 +17,6 @@ Requirements (modules):
 * -4 step 4 - train the segway model
 * -5 step 5 - predict using the segway model
 * -6 step 6 - evaluate the output using segtools
-* -U umapDir - unique mapping directory
-* -S seqDir - sequence directory containing all chromosomes separately
 * -f force - overwrite existing results
 * -a armed - trigger the jobs after writing the scripts
 * -v - print progress information (verbose).
@@ -35,15 +33,13 @@ DO_TRAINSEGWAY=""
 DO_PREDICTSEGWAY=""
 DO_EVALUATE=""
 CLOBBER=""
-UMAPDIR="/share/ClusterShare/biodata/contrib/fabbus/umap/hg19_male/globalmap_k20tok54/"
-SEQDIR="/share/ClusterShare/biodata/contrib/fabbus/encodeDCC/maleByChrom/"
 
 ARMED="FALSE"
 OVERWRITEALL="FALSE"
 
 [ $# -lt 1 ] && echo "$USAGEMSG" >&2 && exit 1
 
-while getopts "123456U:S:afv" opt;
+while getopts "123456afv" opt;
 do
         case ${opt} in
 	1) DO_TRANSFERDATA="TRUE";;
@@ -52,8 +48,6 @@ do
 	4) DO_TRAINSEGWAY="TRUE";;
 	5) DO_PREDICTSEGWAY="TRUE";;
 	6) DO_EVALUATE="TRUE";;
-	U) UMAPDIR="$OPTARG";;
-	S) SEQDIR="$OPTARG";;
         a) ARMED="TRUE";;
         f) OVERWRITEALL="TRUE";;
         v) VERBOSE="--verbose";;
@@ -99,12 +93,14 @@ TRAIN_REGIONS=${SEGWAY_DATA}$(basename ${REGIONS})
 if [ -n "$DO_TRANSFERDATA" ];  then
 	# get all files
 	echo "echo  'collect data tracks from gagri'" > ${SEGWAY_BIN}1_cdata.sh
-
-	for F in ${FILES}; do
+	FILEARR=$(echo $FILES | tr " ," "\n")
+	for F in ${FILEARR}; do
+		FN=${F##*/}
+                FB=${FN%.*}
 	        echo "echo 'datafile ${F}'" >> ${SEGWAY_BIN}1_cdata.sh
-        	echo "smbclient \\\\\\\\gagri\\\\GRIW -A ~/.smbclient -c 'cd ${FILES_SOURCE}/${F}; get ${F}.bam' && mv ${F}.bam ${SEGWAY_DATA}" >> ${SEGWAY_BIN}1_cdata.sh
-	        echo "smbclient \\\\\\\\gagri\\\\GRIW -A ~/.smbclient -c 'cd ${FILES_SOURCE}/${F}; get ${F}.bam.bai' && mv ${F}.bam.bai ${SEGWAY_DATA}" >> ${SEGWAY_BIN}1_cdata.sh
-		echo "smbclient \\\\\\\\gagri\\\\GRIW -A ~/.smbclient -c 'cd ${FILES_SOURCE}/${F}; get ${F}.homer.log' && mv ${F}.homer.log ${SEGWAY_DATA}" >> ${SEGWAY_BIN}1_cdata.sh
+        	echo "smbclient \\\\\\\\gagri\\\\GRIW -A ~/.smbclient -c 'cd ${FILES_SOURCE}${FB}; get ${F}' && mv ${F} ${SEGWAY_DATA}" >> ${SEGWAY_BIN}1_cdata.sh
+	        echo "smbclient \\\\\\\\gagri\\\\GRIW -A ~/.smbclient -c 'cd ${FILES_SOURCE}${FB}; get ${F}.bai' && mv ${F}.bai ${SEGWAY_DATA}" >> ${SEGWAY_BIN}1_cdata.sh
+		echo "smbclient \\\\\\\\gagri\\\\GRIW -A ~/.smbclient -c 'cd ${FILES_SOURCE}${FB}; get ${FB}.homer.log' && mv ${FB}.homer.log ${SEGWAY_DATA}" >> ${SEGWAY_BIN}1_cdata.sh
 
 	done
 
@@ -118,35 +114,48 @@ fi
 ## transform the bam data into bedgraph
 ##
 if [ -n "$DO_CONVERTBAM2BEDGRAPH" ]; then
-	if [ ! -d ${SEQDIR} ] || [ ! -d ${UMAPDIR} ]; then
+	if [ ! -d ${SEQDIR} ] || [ ! -d ${WIGGLER_UMAPDIR} ]; then
 		echo "sequence dir or umap dir is missing/invalid"
 		echo "seqDir: ${SEQDIR}"
-		echo "umapDIR:${UMAPDIR}"
+		echo "umapDIR:${WIGGLER_UMAPDIR}"
 		exit 1
 	fi
 	echo "module load fabbus/wiggler/2.0" > ${SEGWAY_BIN}2_tdata.sh
 	echo "echo 'get chromosome sizes for ${GENOME}'" >> ${SEGWAY_BIN}2_tdata.sh
         echo "fetchChromSizes ${GENOME} > ${SEGWAY_DATA}/${GENOME}.chrom.sizes" >>  ${SEGWAY_BIN}2_tdata.sh
 
-	for F in `ls ${SEGWAY_DATA}*.bam`; do
-        	FN=${F##*/}
-	        FB=${FN%.*}
+	for F in $FILES; do
+		# split replicas
+		REPLICAS=$(echo $F | tr "," "\n")
+		FRAGMENTS=""
+		INPUTS=""
+		REPNAMES=""
+		for REPLICA in $REPLICAS; do
+			[ ! -f ${SEGWAY_DATA}${REPLICA} ] && echo "[ERROR] file not found: ${SEGWAY_DATA}${REPLICA}" && exit 1
+			INPUTS="${INPUTS} -i=${SEGWAY_DATA}${REPLICA}"
+			FRAGSIZE=`grep "Fragment Length Estimate" ${SEGWAY_DATA}${REPLICA%.*}.homer.log | awk '{print $4}'`
+			FRAGMENTS="${FRAGMENTS} -l=${FRAGSIZE}"
+			REPNAMES="${REPNAMES}${REPLICA%.*}-"
+		done
+		# remove last character "-" from REPNAMES
+		REPNAMES="${REPNAMES%?}"
+
+#        	FN=${F##*/}
+#	        FB=${FN%.*}
 
         	if [ ! -f ${SEGWAY_DATA}${FB}.bedgraph.gz ] || [ "$OVERWRITEALL" = "TRUE" ]; then
-                	[ -f ${SEGWAY_QOUT}TrDa-${FB}.out ] && rm ${SEGWAY_QOUT}TrDa-${FB}.out
-			echo '#!/bin/bash' > ${SEGWAY_BIN}tdata${FB}.sh
-	                echo 'echo job_id $JOB_ID startdata $(date)' >> ${SEGWAY_BIN}tdata${FB}.sh
-                	echo "echo convert ${FB}.bam to bedGraph using wiggler" >>  ${SEGWAY_BIN}tdata${FB}.sh
-#        	        echo "genomeCoverageBed -split -bg -ibam ${F} -g ${SEGWAY_DATA}/${GENOME}.chrom.sizes | gzip > ${SEGWAY_DATA}/${FB}.bedgraph.gz" >> ${SEGWAY_BIN}tdata${FB}.sh	
-			# replaced with wiggler
-			FRAGSIZE=`grep "Fragment Length Estimate" ${F%.*}.homer.log | awk '{print $4}'`
-			echo "align2rawsignal -of=wig -i=${F} -s=${SEQDIR} -u=${UMAPDIR} -v=${SEGWAY_QOUT}wiggler-${FN}.log -l=${FRAGSIZE} -k=tukey -w=${WIGGLER_SMOOTHING} -o=${SEGWAY_DATA}${FB}.wig" >> ${SEGWAY_BIN}tdata${FB}.sh 
-			echo "gzip ${SEGWAY_DATA}${FB}.wig" >> ${SEGWAY_BIN}tdata${FB}.sh  
-	                echo 'echo job_id $JOB_ID ending $(date)' >> ${SEGWAY_BIN}tdata${FB}.sh
-			chmod 777 ${SEGWAY_BIN}tdata${FB}.sh
+                	[ -f ${SEGWAY_QOUT}TrDa-${REPNAMES}.out ] && rm ${SEGWAY_QOUT}TrDa-${REPNAMES}.out
+			echo '#!/bin/bash' > ${SEGWAY_BIN}tdata${REPNAMES}.sh
+	                echo 'echo job_id $JOB_ID startdata $(date)' >> ${SEGWAY_BIN}tdata${REPNAMES}.sh
+                	echo "echo convert ${REPNAMES} to bedGraph using wiggler" >>  ${SEGWAY_BIN}tdata${REPNAMES}.sh
+			# wiggler
+			echo "align2rawsignal -of=bg ${INPUTS} ${FRAGMENTS} -s=${SEQDIR} -u=${WIGGLER_UMAPDIR} -v=${SEGWAY_QOUT}wiggler-${REPNAMES}.log -k=tukey -w=${WIGGLER_SMOOTHING} | gzip > ${SEGWAY_DATA}${REPNAMES}.bg.gz" >> ${SEGWAY_BIN}tdata${REPNAMES}.sh 
+#			echo "gzip ${SEGWAY_DATA}${FB}.bg" >> ${SEGWAY_BIN}tdata${REPNAMES}.sh  
+	                echo 'echo job_id $JOB_ID ending $(date)' >> ${SEGWAY_BIN}tdata${REPNAMES}.sh
+			chmod 777 ${SEGWAY_BIN}tdata${REPNAMES}.sh
 
         	        # submit
-                	echo "qsub -V -cwd -l h_rt=01:00:00 -j y -M `whoami`@garvan.unsw.edu.au -S /bin/bash -o ${SEGWAY_QOUT}TrDa-${FB}.out -N TrDa-${FB} ${SEGWAY_BIN}tdata${FB}.sh" >>  ${SEGWAY_BIN}2_tdata.sh
+                	echo "qsub -V -cwd -l h_rt=01:00:00 -j y -M `whoami`@garvan.unsw.edu.au -S /bin/bash -o ${SEGWAY_QOUT}TrDa-${REPNAMES}.out -N TrDa-${REPNAMES} ${SEGWAY_BIN}tdata${REPNAMES}.sh" >>  ${SEGWAY_BIN}2_tdata.sh
 	        fi
 	done
 
@@ -161,21 +170,22 @@ fi
 ## The archieve is then annotated with the bedgraph data  
 ##
 if [ -n "$DO_GENERATEARCHIVE" ]; then
+
+	if [ ! -n "$CHROMSIZES" ];then
+		echo "[ERROR] Chromosome sizes not given: $CHROMSIZES"
+		exit 1
+	fi
 	## load module
 	echo "module load fabbus/segway/1.1.0" > ${SEGWAY_BIN}3_gdata.sh
 	echo "[ -f ${SEGWAY_QOUT}/GnDt-${EXPERIMENT}.out ] && rm ${SEGWAY_QOUT}/GnDt-${EXPERIMENT}.out" >> ${SEGWAY_BIN}3_gdata.sh
 
 
 	echo 'echo job_id $JOB_ID startdata $(date)' > ${SEGWAY_BIN}gdata${EXPERIMENT}.sh
-	echo 'echo get chromosome sizes for ${GENOME}' >> ${SEGWAY_BIN}gdata${EXPERIMENT}.sh
-	[ ! -f ${GENOME}.chrom.sizes ] && echo "fetchChromSizes ${GENOME} > ${SEGWAY_DATA}${GENOME}.chrom.sizes" >> ${SEGWAY_BIN}gdata${EXPERIMENT}.sh
-
 	# genomedata-load call
 	echo "echo '*** create genomedata archive'"  >> ${SEGWAY_BIN}gdata${EXPERIMENT}.sh
-	echo "genomedata-load --sizes -s ${SEGWAY_DATA}${GENOME}.chrom.sizes \\" >> ${SEGWAY_BIN}gdata${EXPERIMENT}.sh
+	echo "genomedata-load --sizes -s ${CHROMSIZES} \\" >> ${SEGWAY_BIN}gdata${EXPERIMENT}.sh
 	# add the -t <ID>=<FILE> sections for all tracks
-
-	for f in $(ls ${SEGWAY_DATA}*.gz ); do
+	for f in $(ls ${SEGWAY_DATA}*.bg.gz ); do
 	        b=$(basename $f)
         	arrIN=(${b//./ })
         echo "-t "${arrIN[0]}=$f" \\" >> ${SEGWAY_BIN}gdata${EXPERIMENT}.sh
